Closes #12, optionally manage svscan service
This commit addresses the topic brought up in #12, where the `svscan`
service is not started. It adds a new attribute that the user can
disable if they don’t want svscan to start or are running on a platform
that doesn’t provide configuration/scripts for their init system to use
daemontools - or are using the source installation.